+++ CHANGELOG.md
@@ -5,7 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [1.1.0]
+## [Unreleased]
+
+### Fixed
+
+- Cannot specify media element [#6](https://github.com/guoyunhe/rabbit-lyrics/issues/6)
+- `data-audio` should be changed to `data-media` in documentation
+
+## [1.1.0] 2018-10-25
 
 ### Added